refactor attributes again
