+++ ffmpeg-gui/src/main/java/attributes/TimeCode.java
@@ -1,4 +1,4 @@
-package entity;
+package attributes;
 
 import lombok.Getter;
