+++ ffmpeg-gui/src/main/java/interface_adapter/add_input_file/AddInputFilePresenter.java
@@ -1,7 +1,7 @@
 package interface_adapter.add_input_file;
 
 import attributes.AudioAttributes;
-import entity.TimeCode;
+import attributes.TimeCode;
 import attributes.VideoAttributes;
 import interface_adapter.ViewManagerModel;
 import interface_adapter.convert_video_file.ConvertVideoFileState;
